requirements.txt
python-dotenv
python-telegram-bot
pandas
xlrd
matplotlib
requests
flask
flask_cors
pymongo